ai-backend/requirements.txt
fastapi
uvicorn[standard]
motor
google-cloud-vision
openai
python-dotenv
PyJWT
python-multipart
PyPDF2==3.0.1
pdf2image==1.16.3